move `ProgramEnvironment` into chalk-solve
