+++ src/coherence.rs
@@ -1,8 +1,8 @@
 use petgraph::prelude::*;
 
 use crate::rust_ir::Program;
-use chalk_ir::ProgramEnvironment;
 use chalk_ir::{self, Identifier, ImplId};
+use chalk_solve::program_environment::ProgramEnvironment;
 use chalk_solve::solve::SolverChoice;
 use failure::Fallible;
 use std::sync::Arc;
