+++ src/coherence/orphan.rs
@@ -5,6 +5,7 @@ use crate::rust_ir::*;
 use chalk_ir::cast::*;
 use chalk_ir::*;
 use chalk_solve::ext::*;
+use chalk_solve::program_environment::ProgramEnvironment;
 use chalk_solve::solve::SolverChoice;
 use failure::Fallible;
