+++ src/rules/wf.rs
@@ -6,6 +6,7 @@ use chalk_ir::fold::shift::Shift;
 use chalk_ir::fold::*;
 use chalk_ir::*;
 use chalk_solve::ext::*;
+use chalk_solve::program_environment::ProgramEnvironment;
 use chalk_solve::solve::SolverChoice;
 use failure::{Fail, Fallible};
 use itertools::Itertools;
